requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
appdirs==1.4.3            # via black, virtualenv
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, pytest
autoflake==1.3
autopep8==1.4.4
backcall==0.1.0           # via ipython
black==19.10b0
cfgv==3.0.0               # via pre-commit
click==7.0                # via black
coverage==4.5.3           # via pytest-cov
decorator==4.4.0          # via ipython, traitlets
distlib==0.3.0            # via virtualenv
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==2.0.0              # via factory-boy
filelock==3.0.12          # via virtualenv
flake8-polyfill==1.0.2    # via pep8-naming
flake8==3.7.8
freezegun==0.3.12
identify==1.4.11          # via pre-commit
importlib-metadata==0.18  # via pluggy, pre-commit, pytest, virtualenv
importmagic==0.1.7
ipython-genutils==0.2.0   # via traitlets
ipython==7.6.1
isort==4.3.21
jedi==0.14.1
mccabe==0.6.1             # via flake8
more-itertools==7.2.0     # via pytest
nodeenv==1.3.5            # via pre-commit
packaging==19.0           # via pytest
parso==0.5.1              # via jedi
pathspec==0.7.0           # via black
pep8-naming==0.8.2
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.12.0            # via pytest
pre-commit==2.1.0
prompt-toolkit==2.0.9     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via autopep8, flake8
pydocstyle==4.0.0
pyflakes==2.1.1           # via autoflake, flake8
pygments==2.4.2           # via ipython
pyparsing==2.4.1.1        # via packaging
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==5.0.1
python-dateutil==2.8.0    # via faker, freezegun
pyyaml==5.3               # via pre-commit
regex==2020.1.8           # via black
rope==0.14.0
six==1.12.0               # via faker, freezegun, packaging, prompt-toolkit, python-dateutil, traitlets, virtualenv
snowballstemmer==1.9.0    # via pydocstyle
text-unidecode==1.2       # via faker
toml==0.10.0              # via black, pre-commit
traitlets==4.3.2          # via ipython
typed-ast==1.4.1          # via black
virtualenv==20.0.4        # via pre-commit
wcwidth==0.1.7            # via prompt-toolkit, pytest
yapf==0.28.0
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
